pyproject.toml
[tool.poetry]
name = "triplea"
version = "0.0.4"
license = "Apache-2.0"
description = "Article Analysis Assistant"
authors = ["Ehsan Bitaraf <bitaraf.e@iums.ac.ir>", "Maryam Jafarpour <maryam.jafarpoor@gmail.com>"]
readme = "README.md"
repository = "https://github.com/EhsanBitaraf/triple-a"
keywords = ["graph", "semantic-scholar", "citation-graph"]
classifiers = [
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
tinydb = "^4.7.1"
pydantic = "^1.10.4"
python-dotenv = "^0.21.1"
xmltodict = "^0.13.0"
networkx = "^3.0"
pandas = "^1.5.3"
nxviz = "^0.7.4"
netwulf = "^0.1.5"
pymongo = "^4.3.3"
py2neo = "^2021.2.3"
ipykernel = "^6.21.2"
ipywidgets = "^8.0.4"
httpx = "^0.25.0"
requests = "^2.31.0"
tomli = "^2.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest-cov = "^4.0.0"


[tool.poetry.scripts]
aaa = "triplea:cli.aaa"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
